models/voxel_models/VoxelFlex-Single-Temperature/requirements.txt
# Core dependencies
torch>=1.10.0
numpy>=1.20.0
pandas>=1.3.0
h5py>=3.1.0
pyyaml>=5.1
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
scipy>=1.6.0

# Enhanced logging dependencies
psutil>=5.8.0  # For memory usage tracking
colorama>=0.4.4  # For colored terminal output (optional)

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
flake8>=3.9.0
black>=21.5b2